# golang/app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an swagger documentation of simple test API task given by tigerhall",
        "title": "Tigerhall test API",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "name": "Tech Support",
            "email": "rg@dooglex.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "tigerhall.dooglex.com",
    "basePath": "/",
    "paths": {
        "/sighting/add": {
            "post": {
                "description": "Create a new sighting of existing tiger",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tiger"
                ],
                "summary": "Create a new sighting of existing tiger",
                "operationId": "create_sighting",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image Upload",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "description": "Request payload",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PayloadAddSighting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseTiger"
                        }
                    }
                }
            }
        },
        "/sighting/show": {
            "post": {
                "description": "show the list of sightings of tigers",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tiger"
                ],
                "summary": "show the list of sightings of tigers",
                "operationId": "show_sighting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page number. Default: 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "description": "Request payload",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.TigerIdModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseShowSighting"
                        }
                    }
                }
            }
        },
        "/tiger/add": {
            "post": {
                "description": "Create a new tiger along with the last seen info",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tiger"
                ],
                "summary": "Create a new tiger along with the last seen info",
                "operationId": "create_tiger",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image Upload",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "description": "Request payload",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PayloadAddNewTiger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseTiger"
                        }
                    }
                }
            }
        },
        "/tiger/show": {
            "post": {
                "description": "show the list of tigers sorted by last seen time",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tiger"
                ],
                "summary": "show the list of tigers sorted by last seen time",
                "operationId": "show_tigers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page number. Default: 1",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseShowTigers"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.PayloadAddNewTiger": {
            "type": "object",
            "properties": {
                "birthday": {
                    "description": "Date of birth of the tiger. Must be in YYYY-MM-DD format.",
                    "type": "string",
                    "example": "2005-12-30"
                },
                "image": {
                    "description": "path to uploaded image.",
                    "type": "string",
                    "example": "path/to/image"
                },
                "last_seen": {
                    "description": "Timestamp when the tiger was last seen. Must be in YYYY-MM-DD format.",
                    "type": "string",
                    "example": "2019-08-01"
                },
                "latitude": {
                    "description": "Last seen Latitude point",
                    "type": "number",
                    "example": 120.51687
                },
                "longitude": {
                    "description": "Last seen Longitude point",
                    "type": "number",
                    "example": 50.894914
                },
                "name": {
                    "description": "Name of the tiger",
                    "type": "string",
                    "example": "tiger08"
                }
            }
        },
        "main.PayloadAddSighting": {
            "type": "object",
            "properties": {
                "image": {
                    "description": "path to uploaded image.",
                    "type": "string",
                    "example": "path/to/image"
                },
                "last_seen": {
                    "description": "Timestamp when the tiger was last seen. Must be in YYYY-MM-DD format.",
                    "type": "string",
                    "example": "2019-08-01"
                },
                "latitude": {
                    "description": "Last seen Latitude point",
                    "type": "number",
                    "example": 120.51687
                },
                "longitude": {
                    "description": "Last seen Longitude point",
                    "type": "number",
                    "example": 50.894914
                },
                "tiger_id": {
                    "description": "id of the tiger",
                    "type": "integer",
                    "example": 7
                }
            }
        },
        "main.ResponseShowSighting": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data field",
                    "type": "object",
                    "properties": {
                        "sighting_data": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.SightingInfo"
                            }
                        },
                        "total_results": {
                            "description": "totals number of results found for the given query",
                            "type": "integer",
                            "example": 20
                        }
                    }
                },
                "status": {
                    "description": "status of the error occurence in the current response",
                    "type": "object",
                    "properties": {
                        "error": {
                            "description": "Whether the current response processed successfully",
                            "type": "boolean",
                            "example": true
                        },
                        "message": {
                            "description": "Error message incase of any error.",
                            "type": "string",
                            "example": "some error information"
                        }
                    }
                }
            }
        },
        "main.ResponseShowTigers": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data field",
                    "type": "object",
                    "properties": {
                        "tiger_data": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ShowTigerModel"
                            }
                        },
                        "total_results": {
                            "description": "totals number of results found for the given query",
                            "type": "integer",
                            "example": 20
                        }
                    }
                },
                "status": {
                    "description": "status of the error occurence in the current response",
                    "type": "object",
                    "properties": {
                        "error": {
                            "description": "Whether the current response processed successfully",
                            "type": "boolean",
                            "example": true
                        },
                        "message": {
                            "description": "Error message incase of any error.",
                            "type": "string",
                            "example": "some error information"
                        }
                    }
                }
            }
        },
        "main.ResponseTiger": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data field",
                    "type": "object",
                    "properties": {
                        "sighting_id": {
                            "description": "primay key for sighting",
                            "type": "integer",
                            "example": 12
                        },
                        "tiger_id": {
                            "description": "id of the tiger",
                            "type": "integer",
                            "example": 7
                        }
                    }
                },
                "status": {
                    "description": "status of the error occurence in the current response",
                    "type": "object",
                    "properties": {
                        "error": {
                            "description": "Whether the current response processed successfully",
                            "type": "boolean",
                            "example": true
                        },
                        "message": {
                            "description": "Error message incase of any error.",
                            "type": "string",
                            "example": "some error information"
                        }
                    }
                }
            }
        },
        "main.ShowTigerModel": {
            "type": "object",
            "properties": {
                "birthday": {
                    "description": "Date of birth of the tiger. Must be in YYYY-MM-DD format.",
                    "type": "string",
                    "example": "2005-12-30"
                },
                "image": {
                    "description": "path to uploaded image.",
                    "type": "string",
                    "example": "path/to/image"
                },
                "last_seen": {
                    "description": "Timestamp when the tiger was last seen. Must be in YYYY-MM-DD format.",
                    "type": "string",
                    "example": "2019-08-01"
                },
                "latitude": {
                    "description": "Last seen Latitude point",
                    "type": "number",
                    "example": 120.51687
                },
                "longitude": {
                    "description": "Last seen Longitude point",
                    "type": "number",
                    "example": 50.894914
                },
                "name": {
                    "description": "Name of the tiger",
                    "type": "string",
                    "example": "tiger08"
                },
                "sighting_id": {
                    "description": "primay key for sighting",
                    "type": "integer",
                    "example": 12
                },
                "tiger_id": {
                    "description": "id of the tiger",
                    "type": "integer",
                    "example": 7
                }
            }
        },
        "main.SightingInfo": {
            "type": "object",
            "properties": {
                "image": {
                    "description": "path to uploaded image.",
                    "type": "string",
                    "example": "path/to/image"
                },
                "last_seen": {
                    "description": "Timestamp when the tiger was last seen. Must be in YYYY-MM-DD format.",
                    "type": "string",
                    "example": "2019-08-01"
                },
                "latitude": {
                    "description": "Last seen Latitude point",
                    "type": "number",
                    "example": 120.51687
                },
                "longitude": {
                    "description": "Last seen Longitude point",
                    "type": "number",
                    "example": 50.894914
                }
            }
        },
        "main.TigerIdModel": {
            "type": "object",
            "properties": {
                "tiger_id": {
                    "description": "id of the tiger",
                    "type": "integer",
                    "example": 7
                }
            }
        }
    }
}